poueeeet!!!

+++ libft/includes/libft.h
@@ -38,7 +38,7 @@ void				*ft_memmove(void *dst, const void *src, size_t len);
 void				*ft_memchr(const void *str, int i, size_t n);
 int					ft_memcmp(const void *str1, const void *str2, size_t n);
 char				*ft_strdup(const char *str);
-size_t				ft_strlen(const char *c);
+int					ft_strlen(const char *c);
 char				*ft_strcpy(char *dst, const char *src);
 char				*ft_strncpy(char *dst, const char *src, size_t n);
 char				*ft_strcat(char *s1, const char *s2);
