+++ libft/srcs/ft_strlen.c
@@ -13,9 +13,9 @@
 #include <string.h>
 #include "libft.h"
 
-size_t			ft_strlen(const char *s)
+int				ft_strlen(const char *s)
 {
-	size_t		cpt;
+	int		cpt;
 
 	cpt = 0;
 	if (s)
